src/requirements.txt
Flask_Cors==4.0.0
imageio==2.34.0
numpy==1.25.2
Pillow==10.2.0
scikit_image==0.22.0
scipy==1.11.2
